move shared blocks to widgets utils
